Fix all Clippy warnings to resolve code quality CI failures
- Remove unnecessary type casts (u32->u32, f32->f32)
- Replace manual div_ceil implementations with built-in method
- Update legacy numeric constants (i32::max_value() -> i32::MAX)
- Fix needless borrows and redundant field names
- Convert single match patterns to if let statements
- Use match for comparison chains
- Change Vec references to slices (&Vec<T> -> &[T])
- Add Default implementation for SimpleChart
- Remove needless return statements
- Handle RefCell references across await points with appropriate allows
- Comment out unused run function in main.rs

+++ src/drawables/plot.rs
@@ -64,9 +64,9 @@ impl RenderListener for PlotRenderer {
                     .iter()
                     .enumerate()
                 {
-                    render_pass.set_vertex_buffer(0 as u32, buffer.slice(..));
+                    render_pass.set_vertex_buffer(0, buffer.slice(..));
                     render_pass.set_vertex_buffer(
-                        1 as u32,
+                        1,
                         ds.data_groups[ds.active_data_group_index].y_buffers[i].slice(..),
                     );
                     log::info!("size: {:?}", buffer.size());
